moved the routes to controller and modified the routes structure
- changed the console log for server listening to a link to make it easier to running.
- upgrading to express 5
- moved the env configs to app to prepare for dependency injection and unit testing.
