+++ package.json
@@ -7,7 +7,7 @@
     "cors": "^2.8.5",
     "dotenv": "^16.0.0",
     "ejs": "^3.1.6",
-    "express": "^4.17.3",
+    "express": "^5.0.0-beta.1",
     "express-session": "^1.17.2",
     "sequelize": "^6.17.0",
     "socket.io": "^4.4.1"
