+++ routes/api/index.api.js
@@ -1,0 +1,6 @@
+const express = require('express');
+const apiRouter = express.Router();
+
+
+
+module.exports = apiRouter;
