+++ routes/web/index.web.js
@@ -1,0 +1,6 @@
+const express = require('express');
+const webRouter = express.Router();
+
+webRouter.use("/", require('./home.router'));
+
+module.exports = webRouter;
